rename: port 및 service를 도메인별로 패키징
