+++ src/main/java/com/ddudu/application/port/in/goal/RetrieveAllGoalsUseCase.java
@@ -1,4 +1,4 @@
-package com.ddudu.application.port.in;
+package com.ddudu.application.port.in.goal;
 
 import com.ddudu.application.domain.goal.dto.response.GoalSummaryResponse;
 import java.util.List;
