+++ src/main/java/com/ddudu/application/service/goal/RetrieveAllGoalsService.java
@@ -1,11 +1,11 @@
-package com.ddudu.application.service;
+package com.ddudu.application.service.goal;
 
 import com.ddudu.application.annotation.UseCase;
 import com.ddudu.application.domain.goal.dto.response.GoalSummaryResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.domain.user.domain.User;
-import com.ddudu.application.port.in.RetrieveAllGoalsUseCase;
-import com.ddudu.application.port.out.GoalLoaderPort;
+import com.ddudu.application.port.in.goal.RetrieveAllGoalsUseCase;
+import com.ddudu.application.port.out.goal.GoalLoaderPort;
 import com.ddudu.application.port.out.UserLoaderPort;
 import java.util.List;
 import java.util.MissingResourceException;
