+++ src/main/java/com/ddudu/application/service/goal/RetrieveGoalService.java
@@ -1,11 +1,11 @@
-package com.ddudu.application.service;
+package com.ddudu.application.service.goal;
 
 import com.ddudu.application.annotation.UseCase;
 import com.ddudu.application.domain.goal.domain.Goal;
 import com.ddudu.application.domain.goal.dto.response.GoalResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
-import com.ddudu.application.port.in.RetrieveGoalUseCase;
-import com.ddudu.application.port.out.GoalLoaderPort;
+import com.ddudu.application.port.in.goal.RetrieveGoalUseCase;
+import com.ddudu.application.port.out.goal.GoalLoaderPort;
 import com.ddudu.application.port.out.UserLoaderPort;
 import java.util.MissingResourceException;
 import lombok.RequiredArgsConstructor;
