+++ src/main/java/com/ddudu/application/service/goal/UpdateGoalService.java
@@ -1,14 +1,14 @@
-package com.ddudu.application.service;
+package com.ddudu.application.service.goal;
 
 import com.ddudu.application.annotation.UseCase;
 import com.ddudu.application.domain.goal.domain.Goal;
-import com.ddudu.application.domain.goal.domain.enums.GoalStatus;
-import com.ddudu.application.domain.goal.dto.request.ChangeGoalStatusRequest;
+import com.ddudu.application.domain.goal.domain.enums.PrivacyType;
+import com.ddudu.application.domain.goal.dto.request.UpdateGoalRequest;
 import com.ddudu.application.domain.goal.dto.response.GoalIdResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
-import com.ddudu.application.port.in.ChangeGoalStatusUseCase;
-import com.ddudu.application.port.out.GoalLoaderPort;
-import com.ddudu.application.port.out.UpdateGoalPort;
+import com.ddudu.application.port.in.goal.UpdateGoalUseCase;
+import com.ddudu.application.port.out.goal.GoalLoaderPort;
+import com.ddudu.application.port.out.goal.UpdateGoalPort;
 import java.util.MissingResourceException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,18 +16,22 @@ import org.springframework.transaction.annotation.Transactional;
 @UseCase
 @RequiredArgsConstructor
 @Transactional
-public class ChangeGoalStatusService implements ChangeGoalStatusUseCase {
+public class UpdateGoalService implements UpdateGoalUseCase {
 
   private final GoalLoaderPort goalLoaderPort;
   private final UpdateGoalPort updateGoalPort;
 
   @Override
-  public GoalIdResponse changeStatus(Long userId, Long id, ChangeGoalStatusRequest request) {
+  public GoalIdResponse update(Long userId, Long id, UpdateGoalRequest request) {
     Goal goal = findGoal(id);
 
     checkAuthority(userId, goal);
 
-    Goal updated = goal.changeStatus(GoalStatus.from(request.status()));
+    Goal updated = goal.applyGoalUpdates(
+        request.name(),
+        request.color(),
+        PrivacyType.from(request.privacyType())
+    );
 
     return GoalIdResponse.from(updateGoalPort.update(updated));
   }
