+++ src/main/java/com/ddudu/presentation/api/controller/GoalController.java
@@ -7,11 +7,11 @@ import com.ddudu.application.domain.goal.dto.response.GoalIdResponse;
 import com.ddudu.application.domain.goal.dto.response.GoalResponse;
 import com.ddudu.application.domain.goal.dto.response.GoalSummaryResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
-import com.ddudu.application.port.in.ChangeGoalStatusUseCase;
-import com.ddudu.application.port.in.CreateGoalUseCase;
-import com.ddudu.application.port.in.RetrieveAllGoalsUseCase;
-import com.ddudu.application.port.in.RetrieveGoalUseCase;
-import com.ddudu.application.port.in.UpdateGoalUseCase;
+import com.ddudu.application.port.in.goal.ChangeGoalStatusUseCase;
+import com.ddudu.application.port.in.goal.CreateGoalUseCase;
+import com.ddudu.application.port.in.goal.RetrieveAllGoalsUseCase;
+import com.ddudu.application.port.in.goal.RetrieveGoalUseCase;
+import com.ddudu.application.port.in.goal.UpdateGoalUseCase;
 import com.ddudu.old.goal.service.GoalService;
 import com.ddudu.presentation.api.annotation.Login;
 import com.ddudu.presentation.api.exception.ForbiddenException;
