+++ src/test/java/com/ddudu/application/service/ChangeGoalStatusServiceTest.java
@@ -8,10 +8,11 @@ import com.ddudu.application.domain.goal.domain.enums.GoalStatus;
 import com.ddudu.application.domain.goal.dto.request.ChangeGoalStatusRequest;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.domain.user.domain.User;
-import com.ddudu.application.port.out.GoalLoaderPort;
-import com.ddudu.application.port.out.SaveGoalPort;
+import com.ddudu.application.port.out.goal.GoalLoaderPort;
+import com.ddudu.application.port.out.goal.SaveGoalPort;
 import com.ddudu.application.port.out.SignUpPort;
 import com.ddudu.application.port.out.UserLoaderPort;
+import com.ddudu.application.service.goal.ChangeGoalStatusService;
 import com.ddudu.fixture.BaseFixture;
 import com.ddudu.fixture.GoalFixture;
 import com.ddudu.fixture.UserFixture;
